+++ src/export.ts
@@ -1,1 +1,2 @@
 export { GUI as Pjax, GUI as default } from './layer/interface/service/gui';
+export { FakeXMLHttpRequest } from './lib/xhr';